+++ include/DeterministicOmegaAutomaton.h
@@ -9,10 +9,10 @@ namespace omalg {
 
   class DeterministicOmegaAutomaton: public virtual OmegaAutomaton {
 	private:
-		std::vector<std::vector<int> > transitionTable;
+		std::vector<std::vector<size_t> > transitionTable;
 		  
 	protected:
-		DeterministicOmegaAutomaton(std::vector<std::vector<int> > theTransitionTable);
+		DeterministicOmegaAutomaton(std::vector<std::vector<size_t> > theTransitionTable);
 
 	public:
 		virtual std::string description() const override;
