+++ src/NondeterministicOmegaAutomaton.cpp
@@ -1,12 +1,12 @@
 #include "NondeterministicOmegaAutomaton.h"
 
 namespace omalg {
-  NondeterministicOmegaAutomaton::NondeterministicOmegaAutomaton(std::vector<std::vector<std::set<int> > > theTransitionRelation)
+  NondeterministicOmegaAutomaton::NondeterministicOmegaAutomaton(std::vector<std::vector<std::set<size_t> > > theTransitionRelation)
     : transitionRelation(theTransitionRelation) {}
 
   std::string NondeterministicOmegaAutomaton::description() const {
     std::string transitionList = "";
-    std::vector<std::vector<std::set<int> > >::const_iterator outerIter;
+    std::vector<std::vector<std::set<size_t> > >::const_iterator outerIter;
     std::vector<std::string> states = this->getStateNames();
     std::vector<std::string> letters = this->getAlphabet();
     for (outerIter = this->transitionRelation.begin(); outerIter != this->transitionRelation.end(); ++outerIter) {
@@ -14,9 +14,9 @@ namespace omalg {
       if (outerIter != this->transitionRelation.begin() && !transitionList.empty() && transitionList.back() != '\n') {
         transitionList += "\n";
       }
-      std::vector<std::set<int> >::const_iterator middleIter;
+      std::vector<std::set<size_t> >::const_iterator middleIter;
       for (middleIter = outerIter->begin(); middleIter != outerIter->end(); ++middleIter) {
-        std::set<int>::const_iterator innerIter;
+        std::set<size_t>::const_iterator innerIter;
         for (innerIter = middleIter->begin(); innerIter != middleIter->end(); ++innerIter) {
           //Add separator but not at beginning of new line
           if (transitionList.back() != '\n' && !transitionList.empty()) {
